refactor!: code is fully splitted into old and new format
