+++ src/__tests__/fileHeader.test.ts
@@ -3,13 +3,13 @@ import { join } from 'path';
 
 import { IOBuffer } from 'iobuffer';
 
-import { TheNewHeader, TheOldHeader } from '../fileHeader';
+import { fileHeader, TheNewHeader, TheOldHeader } from '../fileHeader';
 
 describe('mainHeader parsing test', () => {
   it('m_xyxy.spc', () => {
-    const mxyxy = new TheNewHeader(
+    const mxyxy = fileHeader(
       new IOBuffer(readFileSync(join(__dirname, 'data/m_xyxy.spc'))),
-    );
+    ) as TheNewFileHeader
     expect(mxyxy.parameters.multiFile).toBe(true);
     expect(mxyxy.xUnitsType).toBe('Mass (M/z)');
     expect(mxyxy.zUnitsType).toBe('Minutes');
@@ -17,16 +17,16 @@ describe('mainHeader parsing test', () => {
     expect(mxyxy.memo).toMatch(/^Multiple [^]*X & Y arrays/);
   });
   it('RAMAN.SPC', () => {
-    const raman = new TheNewHeader(
+    const raman = fileHeader(
       new IOBuffer(readFileSync(join(__dirname, 'data/RAMAN.SPC'))),
-    );
+    ) as TheNewFileHeader
     expect(raman.date).toMatch(/1994-08-26T16:45/);
     expect(raman.xyzLabels).toMatch(/Rmn Intensity/);
   });
   it('m_ordz', () => {
-    const mOrdZ = new TheOldHeader(
+    const mOrdZ = fileHeader(
       new IOBuffer(readFileSync(join(__dirname, 'data/m_ordz.spc'))),
-    );
+    ) as TheOldFileHeader
     expect(mOrdZ.fileVersion).toBe(0x4d);
     expect(mOrdZ.memo).toMatch(/^Multiple [^]*ordered Z spacing/);
   });
