+++ src/parse.ts
@@ -1,10 +1,12 @@
 import { IOBuffer } from 'iobuffer';
 
-import { readNewDataBlock, readOldDataBlock, Spectrum } from './dataBlock';
 import { fileHeader, Header, TheNewHeader } from './fileHeader';
 import { LogBlock, readLogBlock } from './logBlock';
+import { newDataBlock, Spectrum } from './newDataBlock';
+import { oldDataBlock } from './oldDataBlock';
 
 export type InputData = ArrayBufferLike | ArrayBufferView | IOBuffer | Buffer;
+
 export interface ParseResult {
   meta: Header;
   spectra: Spectrum[];
@@ -12,31 +14,26 @@ export interface ParseResult {
 }
 
 /**
- * Main header parsing - First 512/256 bytes (new/old format).
- * @param buffer SPC buffer.
- * @return Main header.
- */
-
-/**
  * Parses an SPC file.
  *
  * @param  buffer SPC file buffer.
- * @return Object containing every information contained in the SPC file.
+ * @return JSON-like object with information contained in the SPC file.
  */
 export function parse(buffer: InputData): ParseResult {
   const ioBuffer = new IOBuffer(buffer);
   const meta = fileHeader(ioBuffer);
 
   if (meta instanceof TheNewHeader) {
-    const spectra = readNewDataBlock(ioBuffer, meta);
+    //new format
+    const spectra = newDataBlock(ioBuffer, meta);
     const logs =
       meta.logOffset !== 0 ? readLogBlock(ioBuffer, meta.logOffset) : null;
     return { meta, spectra, logs };
   } else {
-    //the old meta
+    //old format
     return {
       meta,
-      spectra: readOldDataBlock(ioBuffer, meta),
+      spectra: oldDataBlock(ioBuffer, meta),
     };
   }
 }
